update ++
Restructure project components
Now path changes on click
Can reload pages without having to start over
