+++ client/src/components/dashboard/projects/projectComponents.js
@@ -1,5 +1,6 @@
 import React, {useEffect, useState, useRef, createRef} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
+import {useHistory, useLocation } from 'react-router-dom'
 import {book2, settings} from '../../../images/Icon'
 import {useProject} from '../../../contexts/ProjectContext'
 import dayjs from 'dayjs'
@@ -7,7 +8,9 @@ import './css/projectComponent.scss'
 
 function ProjectComponents(props) {
     const users = useSelector((state) => state.users);
-    const { currentProjectComponent, setCurrentProjectComponent} = useProject()
+    const history = useHistory()
+    const location = useLocation()
+    const { currentProjectComponent, setCurrentProjectComponent, currentProject} = useProject()
     const [showSettings, setShowSettings] = useState('')
     const settingsRef = useRef([])
     var relativeTime = require('dayjs/plugin/relativeTime')
@@ -23,6 +26,20 @@ function ProjectComponents(props) {
         }
         // return (username.userName)
     }
+    async function openComponent (e, component) {
+      e.preventDefault()
+      console.log(currentProjectComponent)
+       setCurrentProjectComponent(component)
+      console.log(component)
+      console.log(currentProject)
+      console.log(currentProjectComponent)
+
+      history.push({
+        pathname: '/inn/project-view/' + currentProject._id + '/component/' + currentProjectComponent._id + '/' + currentProjectComponent?.componentPosition,
+        state: currentProjectComponent
+      })
+    }
+
     function openSettings (index, target) {
         console.log(index)
         console.log(target)
@@ -34,7 +51,7 @@ function ProjectComponents(props) {
         <div className="projectComponentContainer">
             {props.components?.slice(0).reverse().map((component, index) => (
                 <div className="rel projectComponentMain" key={index}>
-                    <div className="projectComponent"  ref={settingsRef} onClick={(e) => setCurrentProjectComponent(component)}>
+                    <div className="projectComponent"  ref={settingsRef} onClick={(e) => {openComponent(e, component)}}>
                         <p className="textEffect projectComponentPosition"> {component.componentPosition}</p>
                         <p className="projectComponentName">{component.componentName || 'Unnamed Component'}</p>
                         <img src={book2} className="project-Component-icon" href="Icon-by-Eszter"/>
