+++ client/src/views/css/textTest.scss
@@ -1,0 +1,4 @@
+.testComponent{
+    max-width: 90%;
+    margin: auto;
+}
