+++ client/src/views/index.js
@@ -7,3 +7,6 @@ export { default as LoginPage } from './Login'
 export { default as ForgotPasswordPage } from './ForgotPassword'
 export { default as SignupDetailsPage} from './signupDetailsPage'
 export { default as SinglePost} from './singlePost'
+export { default as TextTest} from './textTest'
+export { default as ProjectView} from './projectView'
+export { default as ProjectComponentView} from './projectComponentView'
